9043/refactor: extract payment calculation logic into dedicated services
