+++ service-api/src/main/java/greencity/service/ubs/calculator/CertificateCalculatorService.java
@@ -1,0 +1,4 @@
+package greencity.service.ubs.calculator;
+
+public interface CertificateCalculatorService {
+}
