+++ service-api/src/main/java/greencity/service/ubs/calculator/PointCalculatorService.java
@@ -1,0 +1,4 @@
+package greencity.service.ubs.calculator;
+
+public interface PointCalculatorService {
+}
