+++ service-api/src/main/java/greencity/service/ubs/user/UbsUserService.java
@@ -1,0 +1,5 @@
+package greencity.service.ubs.user;
+
+public interface UbsUserService {
+
+}
