+++ service/src/main/java/greencity/service/ubs/calculator/CertificateCalculatorServiceImpl.java
@@ -1,0 +1,9 @@
+package greencity.service.ubs.calculator;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class CertificateCalculatorServiceImpl implements CertificateCalculatorService {
+}
